refactor(core): split the 2197-line events module into a package
`domain/events.py` had grown to 2197 lines and 108 event classes covering
every bounded context in the system at once -- lifecycle, invocation, tasks,
health, discovery, auth, operations, administration, enforcement, analysis,
approvals, interceptors. Finding an event meant scrolling; adding one meant
guessing where it belonged; and any change to it touched a file that 141 other
modules import.

Split into thirteen context modules plus `aliases.py`, re-exported from
`__init__.py`. This is a pure move: 112 classes in, 112 classes out, byte-equal
class bodies. The only names that disappeared are leaked imports that were
never part of the intended surface (`dataclass`, `Any`, `uuid`); nothing
imports them.

Three details worth knowing for the next edit:

- The module order is load-bearing, not cosmetic. The four `Provider*` legacy
  aliases are assignments bound to their `McpServer*` counterparts, so
  `aliases.py` must come last. `TestModuleOrderStaysAcyclic` pins this.
- Nothing in the language stops a later edit from adding a class to a submodule
  and forgetting the re-export. That failure is invisible until an importer
  raises at runtime and the event serializer's class map has quietly lost a
  type -- so `test_events_package_surface.py` asserts the `__init__` covers
  every class defined under the package, and that `__all__` matches. Both
  directions were verified to actually fail before being committed.
- `ruff --fix` removed 85 imports that the split made unused. That is the
  monolith's coupling made visible, not new dead code.

Verified: 5518 tests pass, mypy at its 5-error baseline, `lint-imports`
1 kept 0 broken, ruff clean.
