+++ src/mcp_hangar/domain/events/__init__.py
@@ -1,0 +1,274 @@
+# pyright: reportExplicitAny=false
+
+"""Domain events for MCP Hangar.
+
+Events capture important business occurrences and allow decoupled reactions.
+
+This was a single 2197-line module. It is now a package, split along the section
+boundaries the file already carried as comments -- so a reader looking for the
+approval events no longer scrolls past four hundred lines of authorization ones.
+
+Every name is re-exported here, so ``from mcp_hangar.domain.events import X``
+keeps working unchanged. There are 141 such imports across the tree and this
+refactor is deliberately not the place to touch them.
+
+The module order below is definition order and it is load-bearing: the
+``Provider*`` aliases subclass their ``McpServer*`` counterparts, so a module may
+only import from one listed above it.
+"""
+
+from .base import (
+    DomainEvent,
+)
+from .lifecycle import (
+    CircuitBreakerStateChanged,
+    McpServerDegraded,
+    McpServerStarted,
+    McpServerStateChanged,
+    McpServerStopped,
+)
+from .invocation import (
+    ToolInvocationCompleted,
+    ToolInvocationFailed,
+    ToolInvocationRequested,
+)
+from .tasks import (
+    DigestMismatchInTask,
+    TaskCancelled,
+    TaskCompleted,
+    TaskConsentDecided,
+    TaskCreated,
+    TaskFailed,
+    TaskInputRequired,
+)
+from .health import (
+    HealthCheckFailed,
+    HealthCheckPassed,
+    McpServerIdleDetected,
+)
+from .discovery import (
+    DiscoveryCycleCompleted,
+    DiscoverySourceHealthChanged,
+    McpServerApproved,
+    McpServerDiscovered,
+    McpServerDiscoveryConfigChanged,
+    McpServerDiscoveryLost,
+    McpServerQuarantined,
+    ProviderApproved,
+    ProviderDegraded,
+    ProviderDiscovered,
+    ProviderDiscoveryConfigChanged,
+    ProviderDiscoveryLost,
+    ProviderIdleDetected,
+    ProviderQuarantined,
+    ProviderStarted,
+    ProviderStateChanged,
+    ProviderStopped,
+)
+from .auth import (
+    ApiKeyCreated,
+    ApiKeyRevoked,
+    AuthenticationFailed,
+    AuthenticationSucceeded,
+    AuthorizationDenied,
+    AuthorizationGranted,
+    CatalogItemApproved,
+    CatalogItemDeprecated,
+    CatalogItemPublished,
+    CatalogItemRejected,
+    CostReportGenerated,
+    KeyRotated,
+    NamespaceCreated,
+    NamespaceDeleted,
+    PolicyPushRejected,
+    QuotaExceeded,
+    QuotaUpdated,
+    QuotaWarningThresholdReached,
+    RateLimitLockout,
+    RateLimitUnlock,
+    RoleAssigned,
+    RoleRevoked,
+    TenantCreated,
+    TenantReactivated,
+    TenantSuspended,
+)
+from .operations import (
+    BatchCallCompleted,
+    BatchInvocationCompleted,
+    BatchInvocationRequested,
+    ConfigurationReloadFailed,
+    ConfigurationReloadRequested,
+    ConfigurationReloaded,
+    McpServerHotLoaded,
+    McpServerHotUnloaded,
+    McpServerLoadAttempted,
+    McpServerLoadFailed,
+)
+from .administration import (
+    CustomRoleCreated,
+    CustomRoleDeleted,
+    CustomRoleUpdated,
+    McpServerDeregistered,
+    McpServerRegistered,
+    McpServerUpdated,
+    ProviderDeregistered,
+    ProviderRegistered,
+    ProviderUpdated,
+    ToolAccessPolicyCleared,
+    ToolAccessPolicySet,
+)
+from .enforcement import (
+    BehavioralDeviationDetected,
+    BehavioralModeChanged,
+    CapabilityDeclarationMissing,
+    CapabilityViolationDetected,
+    DigestMismatchEvent,
+    EgressBlocked,
+    EgressPolicyCleared,
+    EgressPolicySet,
+    EgressPolicyViolationObserved,
+    McpServerCapabilityQuarantineReleased,
+    McpServerCapabilityQuarantined,
+    ProviderCapabilityQuarantineReleased,
+    ProviderCapabilityQuarantined,
+    ResponseTruncated,
+    ToolSchemaChanged,
+    ToolSchemaDriftDetected,
+)
+from .analysis import (
+    DetectionRuleMatched,
+    EnforcementActionTaken,
+)
+from .approvals import (
+    ToolApprovalDenied,
+    ToolApprovalExpired,
+    ToolApprovalGranted,
+    ToolApprovalRequested,
+    ToolRestored,
+    ToolWithdrawn,
+    ToolWithdrawnRejected,
+)
+from .interceptors import (
+    InterceptorInvoked,
+)
+from .aliases import (
+    ProviderHotLoaded,
+    ProviderHotUnloaded,
+    ProviderLoadAttempted,
+    ProviderLoadFailed,
+)
+
+__all__ = [
+    "ProviderLoadFailed",
+    "ProviderLoadAttempted",
+    "ProviderHotUnloaded",
+    "ProviderHotLoaded",
+    "ApiKeyCreated",
+    "ApiKeyRevoked",
+    "AuthenticationFailed",
+    "AuthenticationSucceeded",
+    "AuthorizationDenied",
+    "AuthorizationGranted",
+    "BatchCallCompleted",
+    "BatchInvocationCompleted",
+    "BatchInvocationRequested",
+    "BehavioralDeviationDetected",
+    "BehavioralModeChanged",
+    "CapabilityDeclarationMissing",
+    "CapabilityViolationDetected",
+    "CatalogItemApproved",
+    "CatalogItemDeprecated",
+    "CatalogItemPublished",
+    "CatalogItemRejected",
+    "CircuitBreakerStateChanged",
+    "ConfigurationReloadFailed",
+    "ConfigurationReloadRequested",
+    "ConfigurationReloaded",
+    "CostReportGenerated",
+    "CustomRoleCreated",
+    "CustomRoleDeleted",
+    "CustomRoleUpdated",
+    "DetectionRuleMatched",
+    "DigestMismatchEvent",
+    "DigestMismatchInTask",
+    "DiscoveryCycleCompleted",
+    "DiscoverySourceHealthChanged",
+    "DomainEvent",
+    "EgressBlocked",
+    "EgressPolicyCleared",
+    "EgressPolicySet",
+    "EgressPolicyViolationObserved",
+    "EnforcementActionTaken",
+    "HealthCheckFailed",
+    "HealthCheckPassed",
+    "InterceptorInvoked",
+    "KeyRotated",
+    "McpServerApproved",
+    "McpServerCapabilityQuarantineReleased",
+    "McpServerCapabilityQuarantined",
+    "McpServerDegraded",
+    "McpServerDeregistered",
+    "McpServerDiscovered",
+    "McpServerDiscoveryConfigChanged",
+    "McpServerDiscoveryLost",
+    "McpServerHotLoaded",
+    "McpServerHotUnloaded",
+    "McpServerIdleDetected",
+    "McpServerLoadAttempted",
+    "McpServerLoadFailed",
+    "McpServerQuarantined",
+    "McpServerRegistered",
+    "McpServerStarted",
+    "McpServerStateChanged",
+    "McpServerStopped",
+    "McpServerUpdated",
+    "NamespaceCreated",
+    "NamespaceDeleted",
+    "PolicyPushRejected",
+    "ProviderApproved",
+    "ProviderCapabilityQuarantineReleased",
+    "ProviderCapabilityQuarantined",
+    "ProviderDegraded",
+    "ProviderDeregistered",
+    "ProviderDiscovered",
+    "ProviderDiscoveryConfigChanged",
+    "ProviderDiscoveryLost",
+    "ProviderIdleDetected",
+    "ProviderQuarantined",
+    "ProviderRegistered",
+    "ProviderStarted",
+    "ProviderStateChanged",
+    "ProviderStopped",
+    "ProviderUpdated",
+    "QuotaExceeded",
+    "QuotaUpdated",
+    "QuotaWarningThresholdReached",
+    "RateLimitLockout",
+    "RateLimitUnlock",
+    "ResponseTruncated",
+    "RoleAssigned",
+    "RoleRevoked",
+    "TaskCancelled",
+    "TaskCompleted",
+    "TaskConsentDecided",
+    "TaskCreated",
+    "TaskFailed",
+    "TaskInputRequired",
+    "TenantCreated",
+    "TenantReactivated",
+    "TenantSuspended",
+    "ToolAccessPolicyCleared",
+    "ToolAccessPolicySet",
+    "ToolApprovalDenied",
+    "ToolApprovalExpired",
+    "ToolApprovalGranted",
+    "ToolApprovalRequested",
+    "ToolInvocationCompleted",
+    "ToolInvocationFailed",
+    "ToolInvocationRequested",
+    "ToolRestored",
+    "ToolSchemaChanged",
+    "ToolSchemaDriftDetected",
+    "ToolWithdrawn",
+    "ToolWithdrawnRejected",
+]
